+++ core/index.js
@@ -283,7 +283,9 @@ function setupProcess(that, lastCallback, cb) {
 }
 
 function setupModules(that, cb) {
-	log.verbose('application configurations:', that.config.getAll());	
+	log.verbose('application configuration path:', that._configPath);
+	log.verbose('application configuration files:', that._configFiles);
+	log.verbose('application configurations:', JSON.stringify(that.config.getAll(), null, 4));	
 	that._module.load(cb);
 }